Move createOrUpdate in pkg/fakerp
